Extract types and helpers
