+++ packages/node/test/integrations/tracing/nest.test.ts
@@ -1,5 +1,5 @@
-import type { InjectableTarget } from '../../../src/integrations/tracing/nest/nest';
-import { isPatched } from '../../../src/integrations/tracing/nest/nest';
+import { isPatched } from '../../../src/integrations/tracing/nest/helpers';
+import type { InjectableTarget } from '../../../src/integrations/tracing/nest/types';
 
 describe('Nest', () => {
   describe('isPatched', () => {
